+++ test/clitest.hs
@@ -230,6 +230,12 @@ main = do
         (_, stdout, stderr) <- runForgeTest "test/contracts/pass/only-deployed-contracts.sol" ["--only-deployed"]
         stderr `shouldNotContain` "CallStack"
         stdout `shouldContain` "[PASS]"
+      it "only-deployed-contracts-force-addr" $ do
+        (_, stdout, stderr) <- runForgeTest "test/contracts/pass/only-deployed-force-addr.sol" ["--only-deployed"]
+        stderr `shouldNotContain` "CallStack"
+        stdout `shouldContain` "[FAIL]"
+        stdout `shouldContain` "[validated]"
+        stderr `shouldNotContain` "not reproducible"
       it "should-fail" $ do
         (_, stdout, stderr) <- runForgeTest "test/contracts/fail/should-fail.sol" []
         stderr `shouldNotContain` "CallStack"